Homepage OG: scope to homepage, bound daily query, null unpriceable volume
Four fixes from review:

1. Move dynamic `generateMetadata` from root `app/layout.tsx` to
   `app/page.tsx` so non-homepage routes (/pool/..., /address-book,
   etc.) don't inherit the cross-chain I/O path on cold cache. Layout
   keeps a static metadata fallback; homepage overrides it. The client
   body of page.tsx moves to page-client.tsx; page.tsx is now a thin
   server wrapper.

2. Add `$since: numeric!` filter (35 days) to a new inline
   HOMEPAGE_OG_DAILY_SNAPSHOTS query so the daily fetch never scans
   the full table. At current scale one page covers every chain; the
   pagination loop is pure headroom for future growth.

3. `rowUsdVolume` now returns `null` for unpriceable rows (both token
   legs without a USD rate). `sumVolumeInWindow` and
   `computeDailyVolumeSeries` skip null rows so an all-unpriceable
   window surfaces as "unavailable" rather than "$0.00".

4. `buildDescription` uses singular "needs" when exactly one pool is
   flagged (matches buildAlt in opengraph-image.tsx).

Also: collapse duplicate tvlSeries.length gate around the chart + label
in opengraph-image.tsx.
