+++ ui-dashboard/src/app/opengraph-image.tsx
@@ -231,20 +231,20 @@ function Card({ data }: { data: HomepageOgData | null }) {
         }}
       >
         {data && data.tvlSeries.length >= 2 ? (
-          <TvlChart series={data.tvlSeries} />
-        ) : null}
-        {data && data.tvlSeries.length >= 2 ? (
-          <span
-            style={{
-              fontSize: 18,
-              letterSpacing: 1.5,
-              textTransform: "uppercase",
-              color: MUTED,
-              alignSelf: "flex-end",
-            }}
-          >
-            Last 30 days
-          </span>
+          <>
+            <TvlChart series={data.tvlSeries} />
+            <span
+              style={{
+                fontSize: 18,
+                letterSpacing: 1.5,
+                textTransform: "uppercase",
+                color: MUTED,
+                alignSelf: "flex-end",
+              }}
+            >
+              Last 30 days
+            </span>
+          </>
         ) : null}
       </div>
     </div>
